+++ src/actors/message.rs
@@ -1,0 +1,13 @@
+use tempfile::NamedTempFile;
+
+use crate::types::{Metadata, Timestamps};
+
+pub type VideoId = String;
+pub type VideoTitle = String;
+
+pub struct DownloadedStream {
+    pub video_id: String,
+    pub file: NamedTempFile,
+    pub metadata: Metadata,
+    pub timestamps: Timestamps,
+}